Refactor Listable
